Moved reference argument parsing to new package
